replace ConnHelper with ClientHelper
